header text size and bbtext tpyes
